Applied all font styles, started links
I've removed the general type scale, i think it's fine if the default sizes apply. There are specific instances that I've adjusted it and that is what i'll keep. I've started adding the link styling but it still needs some work, I still need to add the horizontal rule styling to match the link underlines